apps/api/requirements.txt
fastapi
uvicorn[standard]
celery
redis
fakeredis
sqlalchemy
python-multipart
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
aiofiles
PyPDF2
python-docx
pyyaml
requests
google-generativeai
